+++ expenses/api/currencies.py
@@ -1,16 +1,10 @@
-# common imports
-from datetime import datetime
-from bs4 import BeautifulSoup
-from urllib.request import Request, urlopen
-
 # django drf imports
-from django.conf import settings
-from rest_framework import status, viewsets
+from rest_framework import viewsets
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
 # models import
-from expenses.models import CurrencyConvert, Currency
+from expenses.models import CurrencyConvert
 from expenses.serializers import CurrencyConvertSerializer
 
 from expenses.utils.tools import create_dollar_conversion
@@ -23,5 +17,5 @@ class CurrencyConvertViewSet(viewsets.ModelViewSet):
 
 class CreateDollarConversionView(APIView):
     def post(self, request, *args, **kwargs):
-        data, status = create_dollar_conversion()
-        return Response(data, status=status)
+        data, _status = create_dollar_conversion()
+        return Response(data, status=_status)